Pull import routine into its own file
